Add support for multi-device SyncMessages.
This commit add support for SyncMessages. The client API should be
backwards compatible as no public facing functions where changed.

Added 2 new methods to the client API to expose SyncSent and SyncRead
message types. SyncSent messages are when a linked device sends a
message and SyncRead is when the linked device reads a message. The
client should then handle these accordingly (adding sent message to
database or marking message as read).

Added support for Contact and Group synchronization. When a device
requests contact/group synchronization the local registered contacts are
marshalled and sent as an attachment (SyncContactMessage).

Refactored buildMessage and buildAndSendMessage to use the new
textsecure.Content proto instead of the legacy Body. This was required
to get SyncMessages working properly.